add "supported" item in details box on snap modal
