+++ store_tui/elements/install_modal.py
@@ -1,25 +1,14 @@
-import platform
 from pathlib import Path
 
 import humanize
 from textual.screen import ModalScreen
 from textual.widgets import Footer, Tree
 
+from store_tui.elements.utils import get_platform_architecture
 from store_tui.schemas.snaps.info import ChannelMapItem, InfoResponse
 
 MODAL_CSS_PATH = Path(__file__).parent.parent / "styles" / "snap_modal.tcss"
 PLACEHOLDER_ICON_URL = "https://placehold.co/64/white/black/png?text=?&font=roboto"
-
-
-def get_platform_architecture():
-    machine_arch = platform.machine()
-
-    # remap to snap architectures
-    if machine_arch == "x86_64":
-        return "amd64"
-    elif machine_arch == "aarch64":
-        return "arm64"
-    return machine_arch
 
 
 class InstallModal(ModalScreen):
